Criada a função toggleVisibility para o componente NoteAttachments
