+++ bilhetinhos/src/main/components/notes/Noteboards/NotePreview/index.jsx
@@ -5,7 +5,7 @@ import If from '../../../utils/If'
 import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
 import { setEntireNote } from '../../../../redux/actions/editNoteActions'
-import Slider from 'react-slick'
+import NoteAttachments from '../NoteAttachments/'
 import { Translate } from 'react-translated'
 
 export class NotePreview extends Component {
@@ -68,13 +68,7 @@ export class NotePreview extends Component {
   }
 
   render() {
-    const sliderSettings = {
-      dots: true,
-      infinite: true,
-      speed: 500,
-      slidesToShow: 1,
-      slidesToScroll: 1,
-    }
+    
     return (
       <div className="col-md-6 col-lg-4 d-flex align-items-stretch">
         <div className="card my-3 w-100 drop-material-shadow" id={`note-${this.props.title.replace(/ /g, '')}`}
@@ -119,33 +113,7 @@ export class NotePreview extends Component {
             </ul>
           </div>
           <div className="card-footer">
-            <div className="container">
-              <div className="row">
-                <div className="col">
-                  <button type="button" className="close" data-dismiss="modal" aria-label="Close">
-                    <i className="fas fa-times"></i>
-                  </button>
-                </div>
-                <div className="col">
-                  <div className="p-3">
-                    <Slider {...sliderSettings}>
-                      <div className="d-flex justify-content-center align-items-center">
-                        <img className="slick-img" src="/img/default_cover.png" alt="" style={{ height: '55px', width: '55px' }} />
-                      </div>
-                      <div className="d-flex justify-content-center align-items-center">
-                        <img className="slick-img" src="/img/default_user_profile.png" alt="" style={{ height: '55px', width: '55px' }} />
-                      </div>
-                      <div className="d-flex justify-content-center align-items-center">
-                        <img className="slick-img" src="/img/default_user_profile.png" alt="" style={{ height: '55px', width: '55px' }} />
-                      </div>
-                      <div className="d-flex justify-content-center align-items-center">
-                        <img className="slick-img" src="/img/default_user_profile.png" alt="" style={{ height: '55px', width: '55px' }} />
-                      </div>
-                    </Slider>
-                  </div>
-                </div>
-              </div>
-            </div>
+            <NoteAttachments />
           </div>
         </div>
       </div>
