+++ bilhetinhos/src/main/components/notes/Noteboards/css/NoteAttachments.css
@@ -1,11 +1,11 @@
 
 @keyframes hide {
     from {opacity: 1;}
-    to {opacity: 0;}
+    to {opacity: 0; height: 0}
   }
 
   @keyframes show {
-      from {opacity: 0;}
+      from {opacity: 0; height: 0;}
     to {opacity: 1;}
   }
